requirements.txt
Babel==2.3.4
click==6.6
Django==1.9.6
django-common-helpers==0.8.0
django-cron==0.4.6
django-phonenumber-field==1.1.0
django-rq==0.9.0
django-smsish==1.3.2
Logbook==0.12.5
phonenumberslite==7.3.2
psycopg2==2.6.1
pytz==2016.4
redis==2.10.5
rq==0.6.0
